[front/components/actions/mcp] - refactor: update ToolsList with increased gap
 - Adjust the gap between items in the ToolsList from 2 to 4 for better spacing

[front/components/agent_builder/capabilities] - refactor: rename MCPServerViewsDialog to MCPServerViewsSheet

 - Rename components and types related to MCPServerViewsDialog to Sheet nomenclature for consistency
 - Update imports and state hooks to reflect the new naming of MCPServerViewsSheet

[front/components/agent_builder/capabilities/mcp] - refactor: streamline MCPServerSelection with selected tools state naming

 - Change the name of the state and related props from 'selectedToolsInDialog' to 'selectedToolsInSheet' to match the Sheet component naming
 - Update state and prop usage in MCPServerViewsFooter accordingly

[misc] - refactor: structural improvements to MCP related components

 - Move utility functions from dialogUtils.ts to sheetUtils.ts to align with the renaming of the MCPServerViews component
 - Adjust the utility functions in actionNameUtils.ts to use the new Sheet related variable names

[front/components/agent_builder] - refactor: update constants to reflect new Sheet naming convention

 - Replace CONFIGURATION_DIALOG_PAGE_IDS with TOOLS_SHEET_PAGE_IDS for aligned terminology across the board
