+++ front/components/agent_builder/capabilities/mcp/MCPServerViewsFooter.tsx
@@ -1,7 +1,7 @@
 import { Chip } from "@dust-tt/sparkle";
 import React from "react";
 
-import type { SelectedTool } from "@app/components/agent_builder/capabilities/mcp/MCPServerViewsDialog";
+import type { SelectedTool } from "@app/components/agent_builder/capabilities/mcp/MCPServerViewsSheet";
 import {
   getSelectedToolIcon,
   getSelectedToolLabel,
@@ -9,23 +9,23 @@ import {
 import type { ActionSpecification } from "@app/components/agent_builder/types";
 
 interface MCPServerViewsFooterProps {
-  selectedToolsInDialog: SelectedTool[];
+  selectedToolsInSheet: SelectedTool[];
   dataVisualization?: ActionSpecification | null;
   onRemoveSelectedTool?: (tool: SelectedTool) => void;
 }
 
 export function MCPServerViewsFooter({
-  selectedToolsInDialog,
+  selectedToolsInSheet,
   dataVisualization,
   onRemoveSelectedTool,
 }: MCPServerViewsFooterProps) {
   return (
     <>
-      {selectedToolsInDialog.length > 0 && (
+      {selectedToolsInSheet.length > 0 && (
         <div className="space-y-3">
           <h2 className="text-lg font-semibold">Selected tools</h2>
           <div className="flex flex-wrap gap-2">
-            {selectedToolsInDialog.map((tool, index) => (
+            {selectedToolsInSheet.map((tool, index) => (
               <Chip
                 key={index}
                 icon={getSelectedToolIcon(tool)}
